requirements-alternative.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
keyboard==0.13.5
pydirectinput-rgx==2.1.2